packages/librequant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "librequant"
version = "0.1.0"
description = "Local-first data access for LibreQuant notebooks and strategies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0",
  "pyarrow>=14.0",
  "python-dotenv>=1.0",
  "yfinance>=0.2.40",
  "requests>=2.31",
  "openpyxl>=3.1",
]

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.1"]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.8", "pandas-stubs", "types-requests"]

[tool.hatch.build.targets.wheel]
packages = ["librequant"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "yfinance.*"
ignore_missing_imports = true

# Pandas/psycopg stubs are strict about overloads that do not match dynamic kwargs / psycopg connections.
[[tool.mypy.overrides]]
module = [
    "librequant.data.tabular",
    "librequant.data.postgres",
    "librequant.data.bars",
    "librequant.data.connectors.yfinance",
    "librequant.data.connectors.alpaca",
]
disable_error_code = ["call-overload", "no-any-return", "unused-ignore", "attr-defined"]
